+++ src/BCDeploy.sol
@@ -3,7 +3,7 @@
 pragma solidity ^0.8.24;
 
 import { BCBase } from "./BCBase.sol";
-import { IBCDeployer } from "./interfaces/IBCDeployer.sol";
+import { IBattleChainDeployer } from "@battlechain-contracts/interface/IBattleChainDeployer.sol";
 
 /// @notice Deploy helpers via BattleChainDeployer.
 /// Tracks all deployed addresses for use with BCSafeHarbor.
@@ -11,17 +11,17 @@ abstract contract BCDeploy is BCBase {
     address[] private _deployedContracts;
 
     function bcDeployCreate(bytes memory initCode) internal returns (address deployed) {
-        deployed = IBCDeployer(_bcDeployer()).deployCreate(initCode);
+        deployed = IBattleChainDeployer(_bcDeployer()).deployCreate(initCode);
         _deployedContracts.push(deployed);
     }
 
     function bcDeployCreate2(bytes32 salt, bytes memory initCode) internal returns (address deployed) {
-        deployed = IBCDeployer(_bcDeployer()).deployCreate2(salt, initCode);
+        deployed = IBattleChainDeployer(_bcDeployer()).deployCreate2(salt, initCode);
         _deployedContracts.push(deployed);
     }
 
     function bcDeployCreate3(bytes32 salt, bytes memory initCode) internal returns (address deployed) {
-        deployed = IBCDeployer(_bcDeployer()).deployCreate3(salt, initCode);
+        deployed = IBattleChainDeployer(_bcDeployer()).deployCreate3(salt, initCode);
         _deployedContracts.push(deployed);
     }